# docs/openapi.json
{
    "schemes": [],
    "swagger": "2.0",
    "info": {
        "description": "Task manager API",
        "title": "Task Manager API",
        "contact": {},
        "version": "dev"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/github/link": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Link GitHub account to existing user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "GitHub OAuth link",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Platform for OAuth flow (mobile|web)",
                        "name": "platform",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to Google OAuth consent screen"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/auth/github/login": {
            "get": {
                "description": "Redirect to GitHub OAuth consent screen.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "GitHub OAuth login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Platform for OAuth flow (mobile|web)",
                        "name": "platform",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to Google OAuth consent screen"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/auth/google/link": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Link Google account to existing user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Google OAuth link",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Platform for OAuth flow (mobile|web)",
                        "name": "platform",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to Google OAuth consent screen"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/auth/google/login": {
            "get": {
                "description": "Redirect to Google OAuth consent screen.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Google OAuth login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Platform for OAuth flow (mobile|web)",
                        "name": "platform",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to Google OAuth consent screen"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ErrorCode": {
            "type": "string",
            "enum": [
                "VALIDATION_ERROR",
                "INVALID_REQUEST",
                "UNAUTHORIZED",
                "FORBIDDEN",
                "PAGE_NOT_FOUND",
                "CONFLICT",
                "DATABASE_ERROR",
                "INTERNAL_ERROR",
                "INVALID_TOKEN",
                "MISSING_TOKEN",
                "INVALID_EMAIL"
            ],
            "x-enum-varnames": [
                "CodeValidationError",
                "CodeInvalidRequest",
                "CodeUnauthorized",
                "CodeForbidden",
                "CodeNotFound",
                "CodeConflict",
                "CodeDatabaseError",
                "CodeInternalError",
                "CodeInvalidToken",
                "CodeMissingToken",
                "CodeInvalidEmail"
            ]
        },
        "dto.ErrorData": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/dto.ErrorCode"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.ErrorData"
                },
                "debug": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "trace_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}